khanh(#98)
* tiep tuc cart vs thay doi quantity nhung ap dung ajax bi loi se tim cach fix

* sưa thanh cong update tuy nhien chua ung dung dc ajax do loi nen loai bo kha nang nay

* thuc hien thanh cong buoc thanh toan thieu viec hoan thanh tao oder

* them lai file sau cÃonflict

* xay dung quanh ly don hang tuy con loi va hoat dong so sai

* Them gui otp khi quen, doi mat khau va dang ky

* hieu chinh order

* fix loi merge

---------

+++ project/src/main/java/cart/Cart.java
@@ -3,6 +3,7 @@ package cart;
 import model.Product;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
@@ -33,10 +34,10 @@ public class Cart implements Serializable {
         return total.get();
     }
 
-    public double getTotalPrice(){
+    public BigDecimal getTotalPrice(){
         AtomicReference<Double> total = new AtomicReference<>((double) 0);
         data.values().forEach(p -> total.updateAndGet(v ->  (v + p.getQuantity() * p.getPrice())));
-    return total.get();
+    return BigDecimal.valueOf(total.get());
     }
     public boolean update(int productId, int quantity) {
         CartItem item = data.get(productId);
